requirements.txt
asgiref==3.7.2
attrs==23.1.0
certifi==2023.5.7
charset-normalizer==3.1.0
coreapi==2.3.3
coreschema==0.0.4
Django==4.2.3
django-cors-headers==4.0.0
django-rest-swagger==2.2.0
django-restframework==0.0.1
djangorestframework==3.14.0
et-xmlfile==1.1.0
gcloud==0.18.3
googleapis-common-protos==1.59.1
httplib2==0.22.0
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
jsonschema==4.17.3
jws==0.1.3
MarkupSafe==2.1.2
numpy==1.25.0
oauth2client==4.1.3
openapi==1.1.0
openapi-codec==1.3.2
openpyxl==3.1.2
pandas==2.0.3
protobuf==4.23.4
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycryptodome==3.18.0
pyparsing==3.1.0
Pyrebase4==4.7.1
pyrsistent==0.19.3
python-dateutil==2.8.2
python-jwt==2.0.1
pytz==2023.3
PyYAML==6.0
requests==2.29.0
requests-toolbelt==0.10.1
rsa==4.9
simplejson==3.19.1
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.7.1
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.16
